Add black and flake8 formatting

+++ modelling/disease_selector.py
@@ -1,7 +1,10 @@
 from typing import Tuple
 from pandas.core.frame import DataFrame
 
-def get_disease_data(disease_to_select: str, disease_target: DataFrame) -> Tuple[DataFrame, DataFrame, DataFrame]:
+
+def get_disease_data(
+    disease_to_select: str, disease_target: DataFrame
+) -> Tuple[DataFrame, DataFrame, DataFrame]:
     """
     Creates a single disease subset dataframe from the full metagenomic dataframe.
     Args:
@@ -9,17 +12,25 @@ def get_disease_data(disease_to_select: str, disease_target: DataFrame) -> Tuple
         abundance_data: (DataFrame): Dataframe of the targeted disease subset
 
     Returns:
-        Tuple [DataFrame, DataFrame, DataFrame]: 3 dataframes returning subsets of the dataset containing control samples, disease samples, and both, with a boolean value in the disease column.
+        Tuple [DataFrame, DataFrame, DataFrame]: 3 dataframes returning subsets
+        of the dataset containing control samples, disease samples, and both,
+        with a boolean (int) value in the disease column.
     """
 
     disease = disease_target["disease"]
 
     # Create subset of controls and target diseases
-    combined_disease_control = disease_target[(disease == "n") | (disease == "nd") | (disease == disease_to_select)]
+    combined_disease_control = disease_target[
+        (disease == "n") | (disease == "nd") | (disease == disease_to_select)
+    ]
     control = disease_target[(disease == "n") | (disease == "nd")]
     disease_target = disease_target[disease == disease_to_select]
 
     # transform target variable disease into numeric data
-    combined_disease_control["disease"] = combined_disease_control["disease"].replace(["n", "nd"], 0)
-    combined_disease_control["disease"] = combined_disease_control["disease"].replace(disease_to_select, 1)
+    combined_disease_control["disease"] = combined_disease_control["disease"].replace(
+        ["n", "nd"], 0
+    )
+    combined_disease_control["disease"] = combined_disease_control["disease"].replace(
+        disease_to_select, 1
+    )
     return combined_disease_control, disease_target, control
